+++ src/taskArray.js
@@ -1,0 +1,17 @@
+export default [
+  {
+    description: 'wash the dishes',
+    completed: false,
+    index: 0,
+  },
+  {
+    description: 'learn code',
+    completed: false,
+    index: 1,
+  },
+  {
+    description: 'complete todo list project',
+    completed: false,
+    index: 2,
+  },
+];
